refactor: remove duplicated public files
