--- public/.firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "matheus-35023"
-  }
-}
