--- public/firebase.json
@@ -1,20 +1,0 @@
-{
-  "functions": {
-    "source": "functions",
-    "predeploy": []
-  },
-  "hosting": {
-    "public": "public",
-    "ignore": [
-      "firebase.json",
-      "**/.*",
-      "**/node_modules/**"
-    ],
-    "rewrites": [
-      {
-        "source": "**",
-        "destination": "/index.html"
-      }
-    ]
-  }
-}
